document upload controller and service done

+++ src/SoowGoodWeb.Domain/Models/DocumentsAttachment.cs
@@ -14,7 +14,7 @@ namespace SoowGoodWeb.Models
         public string? OriginalFileName { get; set; }
         public string? Path { get; set; }
         public EntityType? EntityType { get; set; }
-        public int? EntityId { get; set; }
+        public long? EntityId { get; set; }
         public AttachmentType? AttachmentType { get; set; }
     }
 }
